refactor(settings): move vector-db controls to embeddings tab
The Context tab owned embedding/vector-DB controls that belong with the
dedicated Embeddings tab, and carried a second, older copy of the
embedding-health + rebuild flow the Embeddings tab already implements via
useEmbeddingRebuild/EmbeddingHealthAlert. That split the Context/Embeddings
boundary and buried the app's most destructive actions under retrieval tuning.

Relocate StorageStatsCard, DatabaseManagementCard, EmbeddingLimitsConfig, and
EmbeddingIndexControls into embedding-settings.tsx, along with the storage/cache
stats loader and the remove-duplicates / clear-chat / clear-all handlers and
their confirm dialog. Maintenance actions refresh both storage and dimension
stats. The destructive trio stays grouped in DatabaseManagementCard (already a
trash-iconed card), placed after generation config.

Context is left with conversation context, prompt budget, grounding, web
search, retrieval, and files/chunking only — and its duplicated health/rebuild
block is dropped (the Embeddings tab's is authoritative).

Update the settings registry: the moved ids now report tab "embeddings".
settings-page resolves a deep-linked ?focus id's tab through the registry, so
old ?tab=context links to relocated controls land on the embeddings tab.
