+++ src/features/model/components/embedding-settings.tsx
@@ -1,14 +1,16 @@
 import { useStorage } from "@plasmohq/storage/hook"
-import { useCallback, useEffect, useMemo, useState } from "react"
+import { useCallback, useEffect, useMemo, useRef, useState } from "react"
 import { useTranslation } from "react-i18next"
-import { SectionStack } from "@/components/layout"
-import { StatusAlert } from "@/components/settings"
+import { SectionStack, TwoColumnGrid } from "@/components/layout"
+import { ConfirmActionDialog, StatusAlert } from "@/components/settings"
 import { DataMigrationSettings } from "@/features/knowledge/components/data-migration-settings"
 import { FeedbackSettings } from "@/features/knowledge/components/feedback-settings"
 import { useEmbeddingDimensionStats } from "@/features/model/hooks/use-embedding-dimension-stats"
 import { useEmbeddingModelCheck } from "@/features/model/hooks/use-embedding-model-check"
 import { useEmbeddingRebuild } from "@/features/model/hooks/use-embedding-rebuild"
 import { useProviderModels } from "@/features/model/hooks/use-provider-models"
+import { useConfirmAction } from "@/hooks/use-confirm-action"
+import { useToast } from "@/hooks/use-toast"
 import {
   DEFAULT_EMBEDDING_CONFIG,
   DEFAULT_EMBEDDING_MODEL,
@@ -16,19 +18,30 @@ import {
   type EmbeddingConfig,
   STORAGE_KEYS
 } from "@/lib/constants"
+import { getCacheStats } from "@/lib/embeddings/embedding-client"
 import {
   isLikelyEmbeddingModelName,
   recommendedEmbeddingBaseSet
 } from "@/lib/embeddings/model-name-filter"
+import {
+  clearAllVectors,
+  getStorageStats,
+  removeDuplicateVectors
+} from "@/lib/embeddings/vector-store"
+import { logger } from "@/lib/logger"
 import { AlertTriangle, RefreshCw } from "@/lib/lucide-icon"
 import { plasmoGlobalStorage } from "@/lib/plasmo-global-storage"
 
+import { DatabaseManagementCard } from "./embedding-config/database-management-card"
 import { EmbeddingGenerationConfig } from "./embedding-config/embedding-generation-config"
 import { EmbeddingHealthAlert } from "./embedding-config/embedding-health-alert"
+import { EmbeddingLimitsConfig } from "./embedding-config/embedding-limits-config"
 import { EmbeddingModelSelector } from "./embedding-config/embedding-model-selector"
 import { EmbeddingRebuildDialogs } from "./embedding-config/embedding-rebuild-dialogs"
 import { EmbeddingTestGeneration } from "./embedding-config/embedding-test-generation"
 import { EmbeddingTestSearch } from "./embedding-config/embedding-test-search"
+import { StorageStatsCard } from "./embedding-config/storage-stats-card"
+import { EmbeddingIndexControls } from "./embedding-index-controls"
 
 /**
  * Orchestrator for the Embeddings settings screen.
@@ -46,6 +59,7 @@ import { EmbeddingTestSearch } from "./embedding-config/embedding-test-search"
  */
 export const EmbeddingSettings = () => {
   const { t } = useTranslation()
+  const { toast } = useToast()
 
   const [selectedModel, setSelectedModel] = useStorage<string>(
     {
@@ -150,6 +164,169 @@ export const EmbeddingSettings = () => {
     resolveProviderForModel
   })
 
+  // Vector-store stats + maintenance, relocated here from the Context tab so
+  // every embedding/vector-DB control lives on one screen. Dimension stats
+  // come from `useEmbeddingDimensionStats`; storage/cache totals are loaded
+  // locally and refreshed after any destructive action.
+  const [storageStats, setStorageStats] = useState<{
+    totalVectors: number
+    totalSizeMB: number
+    byType: Record<string, number>
+  } | null>(null)
+  const [cacheStats, setCacheStats] = useState<{
+    size: number
+    maxSize: number
+  } | null>(null)
+  const [isCleaning, setIsCleaning] = useState(false)
+  const [confirmAction, setConfirmAction] = useState<
+    "removeDuplicates" | "clearChat" | "clearAll" | null
+  >(null)
+  const confirmDialog = useConfirmAction()
+  const isLoadingStatsRef = useRef(false)
+
+  const loadStorageStats = useCallback(async () => {
+    if (isLoadingStatsRef.current) return
+    isLoadingStatsRef.current = true
+    try {
+      const [statsResult] = await Promise.allSettled([getStorageStats()])
+      if (statsResult.status === "fulfilled") setStorageStats(statsResult.value)
+      setCacheStats(getCacheStats())
+    } catch (error) {
+      logger.error("Failed to load storage stats", "EmbeddingSettings", {
+        error
+      })
+    } finally {
+      isLoadingStatsRef.current = false
+    }
+  }, [])
+
+  useEffect(() => {
+    loadStorageStats()
+    const interval = setInterval(() => {
+      if (document.visibilityState === "visible") loadStorageStats()
+    }, 10000)
+    return () => clearInterval(interval)
+  }, [loadStorageStats])
+
+  const refreshAfterMaintenance = useCallback(async () => {
+    await Promise.all([loadStorageStats(), refreshDimensionStats()])
+  }, [loadStorageStats, refreshDimensionStats])
+
+  const handleRemoveDuplicates = useCallback(async () => {
+    setIsCleaning(true)
+    try {
+      const { deleted, kept } = await removeDuplicateVectors()
+      toast({
+        title: t(
+          "model.embedding_config.database_management.remove_duplicates_success",
+          { deleted, kept }
+        )
+      })
+      await refreshAfterMaintenance()
+    } catch (error) {
+      logger.error("Failed to remove duplicates", "EmbeddingSettings", {
+        error
+      })
+      toast({
+        title: t(
+          "model.embedding_config.database_management.remove_duplicates_error"
+        ),
+        variant: "destructive"
+      })
+    } finally {
+      setIsCleaning(false)
+    }
+  }, [refreshAfterMaintenance, t, toast])
+
+  const handleClearChatVectors = useCallback(async () => {
+    setIsCleaning(true)
+    try {
+      const deleted = await clearAllVectors("chat")
+      toast({
+        title: t(
+          "model.embedding_config.database_management.clear_chat_success",
+          { count: deleted }
+        )
+      })
+      await refreshAfterMaintenance()
+    } catch (error) {
+      logger.error("Failed to clear chat vectors", "EmbeddingSettings", {
+        error
+      })
+      toast({
+        title: t("model.embedding_config.database_management.clear_chat_error"),
+        variant: "destructive"
+      })
+    } finally {
+      setIsCleaning(false)
+    }
+  }, [refreshAfterMaintenance, t, toast])
+
+  const handleClearAllVectors = useCallback(async () => {
+    setIsCleaning(true)
+    try {
+      await clearAllVectors()
+      toast({
+        title: t("model.embedding_config.database_management.clear_all_success")
+      })
+      await refreshAfterMaintenance()
+    } catch (error) {
+      logger.error("Failed to clear all vectors", "EmbeddingSettings", {
+        error
+      })
+      toast({
+        title: t("model.embedding_config.database_management.clear_all_error"),
+        variant: "destructive"
+      })
+    } finally {
+      setIsCleaning(false)
+    }
+  }, [refreshAfterMaintenance, t, toast])
+
+  const openConfirm = useCallback(
+    (action: "removeDuplicates" | "clearChat" | "clearAll") => {
+      setConfirmAction(action)
+      confirmDialog.openDialog()
+    },
+    [confirmDialog]
+  )
+
+  const closeConfirm = useCallback(() => {
+    confirmDialog.closeDialog()
+    setConfirmAction(null)
+  }, [confirmDialog])
+
+  const confirmConfig = (() => {
+    switch (confirmAction) {
+      case "removeDuplicates":
+        return {
+          title: t(
+            "model.embedding_config.database_management.remove_duplicates_confirm"
+          ),
+          confirmLabel: t("model.embedding_config.remove_duplicates_button"),
+          onConfirm: handleRemoveDuplicates
+        }
+      case "clearChat":
+        return {
+          title: t(
+            "model.embedding_config.database_management.clear_chat_confirm"
+          ),
+          confirmLabel: t("model.embedding_config.clear_chat_button"),
+          onConfirm: handleClearChatVectors
+        }
+      case "clearAll":
+        return {
+          title: t(
+            "model.embedding_config.database_management.clear_all_confirm"
+          ),
+          confirmLabel: t("model.embedding_config.clear_all_button"),
+          onConfirm: handleClearAllVectors
+        }
+      default:
+        return null
+    }
+  })()
+
   // Dialog open-state lives here so it can drive the dialog component.
   const [confirmRebuildOpen, setConfirmRebuildOpen] = useState(false)
   const [modelChangeOpen, setModelChangeOpen] = useState(false)
@@ -228,8 +405,39 @@ export const EmbeddingSettings = () => {
         config={config || DEFAULT_EMBEDDING_CONFIG}
         updateConfig={updateConfig}
       />
+      {storageStats && (
+        <StorageStatsCard storageStats={storageStats} cacheStats={cacheStats} />
+      )}
+      <TwoColumnGrid>
+        <DatabaseManagementCard
+          onRemoveDuplicates={() => openConfirm("removeDuplicates")}
+          onClearChat={() => openConfirm("clearChat")}
+          onClearAll={() => openConfirm("clearAll")}
+          isCleaning={isCleaning || isRebuilding}
+          hasVectors={!!storageStats?.totalVectors}
+          hasChatVectors={!!storageStats?.byType?.chat}
+        />
+        <EmbeddingLimitsConfig
+          config={config || DEFAULT_EMBEDDING_CONFIG}
+          updateConfig={updateConfig}
+        />
+      </TwoColumnGrid>
+      <EmbeddingIndexControls />
       <FeedbackSettings />
       <DataMigrationSettings />
+      <ConfirmActionDialog
+        open={confirmDialog.open}
+        onOpenChange={(next) => {
+          if (!next) closeConfirm()
+        }}
+        title={confirmConfig?.title || ""}
+        confirmLabel={confirmConfig?.confirmLabel || t("common.save")}
+        onConfirm={async () => {
+          if (!confirmConfig) return
+          closeConfirm()
+          await confirmConfig.onConfirm()
+        }}
+      />
       <EmbeddingRebuildDialogs
         confirmRebuildOpen={confirmRebuildOpen}
         onConfirmRebuildOpenChange={setConfirmRebuildOpen}
